set up an object for the board and wrote welcome
It took me a while to get the line break \n syntax, and I am still unsure about this string because in code pen, it is printed in double quotes and some of the words and characters are different colors.

I don't think it likes !, Game, For, and 1. and 2. within a string.